classifier/requirements.txt
boto3==1.14.28
botocore==1.17.28
certifi==2020.6.20
chardet==3.0.4
docutils==0.15.2
future==0.18.2
idna==2.10
jmespath==0.10.0
numpy==1.19.1
Pillow==7.2.0
python-dateutil==2.8.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
torch==1.5.1
torchvision==0.6.1
urllib3==1.25.10
